requirements.txt
streamlit
fastapi 
requests 
bcrypt 
PyJWT 
pydantic